# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /verify:
    get:
      operationId: getVerifyReport
      tags: [verify]
      summary: Data health reconciliation report
      description: >
        Compares three sources for the same fiscal year — xlsx as ingested,
        live Google Sheets read now, and the database — reporting row counts,
        amount sums, distinct invoices/customers, by-group and by-head
        breakdowns, deltas over 0.5 percent, and live rows missing from the
        database.
      parameters:
        - name: fy
          in: query
          required: false
          description: Fiscal year, e.g. 2026-27 (default).
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}$"
      responses:
        "200":
          description: Verification report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyReport"
        "400":
          description: Unknown fiscal year
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Report could not be built
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /verify/backfill:
    post:
      operationId: runVerifyBackfill
      tags: [verify]
      summary: Backfill rows missing from the database
      description: >
        Re-reads the live register for the fiscal year and inserts any lines
        whose uid is not yet in the database. Idempotent.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyBackfillRequest"
      responses:
        "200":
          description: Backfill result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyBackfillResponse"
        "400":
          description: Unknown fiscal year
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Backfill failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
      required:
        - error
    VerifyBreakdown:
      type: object
      properties:
        key:
          type: string
          description: Canonical group or head name.
        amount:
          type: number
          description: Amount in INR.
      required:
        - key
        - amount
    VerifySourceAggregates:
      type: object
      properties:
        rows:
          type: integer
          description: Row count for the fiscal year.
        amount:
          type: number
          description: SUM(amount) in INR.
        invoices:
          type: integer
          description: Distinct invoice numbers.
        customers:
          type: integer
          description: Distinct customers.
        byGroup:
          type: array
          items:
            $ref: "#/components/schemas/VerifyBreakdown"
        byHead:
          type: array
          items:
            $ref: "#/components/schemas/VerifyBreakdown"
      required:
        - rows
        - amount
        - invoices
        - customers
        - byGroup
        - byHead
    VerifyDelta:
      type: object
      properties:
        metric:
          type: string
          description: Metric compared (rows, amount, invoices, customers).
        a:
          type: number
          description: Value from the first source.
        b:
          type: number
          description: Value from the second source.
        deltaPct:
          type: number
          description: Absolute percentage difference.
        flagged:
          type: boolean
          description: True when the delta exceeds 0.5 percent.
      required:
        - metric
        - a
        - b
        - deltaPct
        - flagged
    VerifyComparison:
      type: object
      properties:
        label:
          type: string
          description: Which two sources are compared.
        deltas:
          type: array
          items:
            $ref: "#/components/schemas/VerifyDelta"
        flagged:
          type: boolean
          description: True when any delta is over threshold.
      required:
        - label
        - deltas
        - flagged
    VerifyMissingRow:
      type: object
      properties:
        invoiceNo:
          type: string
          nullable: true
        code:
          type: string
        qty:
          type: number
          nullable: true
        amount:
          type: number
        monthLabel:
          type: string
          nullable: true
        customer:
          type: string
          nullable: true
      required:
        - code
        - amount
    VerifyReport:
      type: object
      properties:
        fy:
          type: string
          description: Fiscal year verified.
        generatedAt:
          type: string
          description: ISO timestamp of report generation.
        sources:
          type: object
          properties:
            xlsx:
              $ref: "#/components/schemas/VerifySourceAggregates"
            sheets:
              $ref: "#/components/schemas/VerifySourceAggregates"
            db:
              $ref: "#/components/schemas/VerifySourceAggregates"
          required:
            - xlsx
            - sheets
            - db
        comparisons:
          type: array
          items:
            $ref: "#/components/schemas/VerifyComparison"
        missingFromDb:
          type: object
          properties:
            count:
              type: integer
              description: Live rows whose uid is not in the database.
            sample:
              type: array
              description: Up to 50 example missing rows.
              items:
                $ref: "#/components/schemas/VerifyMissingRow"
          required:
            - count
            - sample
        healthy:
          type: boolean
          description: True when nothing is missing and no delta is flagged.
      required:
        - fy
        - generatedAt
        - sources
        - comparisons
        - missingFromDb
        - healthy
    VerifyBackfillRequest:
      type: object
      properties:
        fy:
          type: string
          description: Fiscal year, e.g. 2026-27 (default).
    VerifyBackfillResponse:
      type: object
      properties:
        fy:
          type: string
        rowsRead:
          type: integer
          description: Live rows read for the fiscal year.
        inserted:
          type: integer
          description: Rows inserted (0 when already in sync).
      required:
        - fy
        - rowsRead
        - inserted
...